move CartRuleSettings ns
